aider: Set total volume size and block size in the `statfs` method.
# Aider chat conversation:

USER: when i am mounting the volume with the .mount() function, it seems to work but I am unable to set the total volume and created at date. is there a way to set that value?

ASSISTANT: The total volume size and creation date are typically managed by the file system itself and not directly settable via the FUSE API. However, you can provide an illusion of these properties by implementing the `statfs` method in your FUSE file system.

The `statfs` method is used to retrieve file system statistics. These statistics include data such as total blocks (`f_blocks`), free blocks (`f_bfree`), and block size (`f_bsize`), which can be used to calculate the total volume size.

As for the creation date, FUSE does not provide a direct way to set this. File creation dates are typically managed by the underlying file system and are not modifiable via the FUSE API. However, you can maintain this information within your file system implementation and return it when file attributes are requested.

In your provided code, the `statfs` method is already implemented and sets the total blocks and block size for the Windows platform. You can modify this method to set these values for other platforms as well.

Here is an example of how you can modify the `statfs` method to set the total volume size:

```java
src/main/java/ru/serce/jnrfuse/examples/MemoryFS.java
<<<<<<< HEAD
    @Override
    public int statfs(String path, Statvfs stbuf) {
        System.out.println("statfs() called with arguments: path = " + path);// + ", stbuf = " + stbuf);
        if (Platform.getNativePlatform().getOS() == WINDOWS) {
            // statfs needs to be implemented on Windows in order to allow for copying
            // data from other devices because winfsp calculates the volume size based
            // on the statvfs call.
            // see https://github.com/billziss-gh/winfsp/blob/14e6b402fe3360fdebcc78868de8df27622b565f/src/dll/fuse/fuse_intf.c#L654
            if ("/".equals(path)) {
                stbuf.f_blocks.set(1024 * 1024); // total data blocks in file system
                stbuf.f_frsize.set(1024);        // fs block size
                stbuf.f_bfree.set(1024 * 1024);  // free blocks in fs
            }
        }
        return super.statfs(path, stbuf);
    }
=======
    @Override
    public int statfs(String path, Statvfs stbuf) {
        System.out.println("statfs() called with arguments: path = " + path);// + ", stbuf = " + stbuf);
        // Set total data blocks in file system, fs block size, and free blocks in fs
        stbuf.f_blocks.set(1024 * 1024);
        stbuf.f_frsize.set(1024);
        stbuf.f_bfree.set(1024 * 1024);
        return 0;
    }
>>>>>>> updated
```

Please note that this change will set the total volume size to approximately 1GB (1024 * 1024 blocks of size 1024 bytes each), regardless of the actual available space in the underlying storage. Adjust these values as needed to reflect the actual size of your file system.

As for the creation date, as I mentioned earlier, it's not directly supported by FUSE. You might need to maintain this information in your own data structures and provide it when necessary.