aider: Set total, free, and available blocks to a large value in `statfs` method to address Finder's space issue.
# Aider chat conversation:

ASSISTANT: To address the issue with Finder on macOS reporting that there is not enough free space to drag and drop files into the mounted virtual drive, we need to modify the `statfs` method in your `MemoryFS` class to return a large enough value for the available blocks (`f_bavail`) and the total blocks (`f_blocks`). This will make the Finder believe that there is sufficient space available on the filesystem.

Here is the diff for the changes to be made in the `MemoryFS.java` file:

```diff
--- /Users/btcmac/jnr-fuse/src/main/java/ru/serce/jnrfuse/examples/MemoryFS.java
+++ /Users/btcmac/jnr-fuse/src/main/java/ru/serce/jnrfuse/examples/MemoryFS.java
@@ ... @@
     @Override
     public int statfs(String path, Statvfs stbuf) {
         if (Platform.getNativePlatform().getOS() == WINDOWS) {
             // statfs needs to be implemented on Windows in order to allow for copying
             // data from other devices because winfsp calculates the volume size based
             // on the statvfs call.
             // see https://github.com/billziss-gh/winfsp/blob/14e6b402fe3360fdebcc78868de8df27622b565f/src/dll/fuse/fuse_intf.c#L654
             if ("/".equals(path)) {
-                stbuf.f_blocks.set(1024 * 1024); // total data blocks in file system
+                stbuf.f_blocks.set(1024 * 1024 * 1024); // total data blocks in file system
                 stbuf.f_frsize.set(1024);        // fs block size
-                stbuf.f_bfree.set(1024 * 1024);  // free blocks in fs
+                stbuf.f_bfree.set(1024 * 1024 * 1024);  // free blocks in fs
+                stbuf.f_bavail.set(1024 * 1024 * 1024); // free blocks available to unprivileged user
             }
         }
         //called all the time...too much spam on printouts
         //System.out.println("statfs() called with arguments: path = " + path);// + ", stbuf = " + stbuf);
         return super.statfs(path, stbuf);
     }
@@ ... @@
```

This change sets the total number of blocks, the number of free blocks, and the number of free blocks available to unprivileged users to a large value (1 billion blocks of 1KB each), which should be more than enough to allow file operations from Finder. Please apply this patch to your `MemoryFS.java` file and test the drag and drop functionality again.